requirements.txt
Flask==1.1.1
google-cloud-firestore==1.6.2
google-cloud-monitoring==0.34.0
pygithub==1.50
python-dateutil==2.8.1
requests==2.23.0
